+++ lib/use-digital-stage/useMediasoup/index.tsx
@@ -1,5 +1,6 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import debug from 'debug';
+import omit from 'lodash/omit';
 import {
   useAudioConsumers,
   useAudioProducers,
@@ -12,13 +13,15 @@ import useMediasoupTransport from './useMediasoupTransport';
 import allActions from '../redux/actions';
 import { LocalConsumer, LocalProducer, RemoteAudioProducer, RemoteVideoProducer } from '../types';
 import { getAudioTracks, getVideoTracks } from './util';
-import omit from 'lodash/omit';
 
-const d = debug('useMediasoup');
-const trace = d.extend('info');
-const err = d.extend('err');
+const sampleRate: number | undefined = process.env.NEXT_PUBLIC_FIXED_SAMPLERATE
+  ? parseInt(process.env.NEXT_PUBLIC_FIXED_SAMPLERATE, 10)
+  : undefined;
 
-const debugEffect = debug('useEffect:useMediasoup');
+const report = debug('useMediasoup');
+const reportError = report.extend('error');
+
+const reportEffect = debug('useEffect:useMediasoup');
 
 function isAudioProducer(
   producer: RemoteVideoProducer | RemoteAudioProducer
@@ -39,8 +42,28 @@ const MediasoupProvider = (props: {
   const [sendVideo, setSendVideo] = useState<boolean>(false);
   const [receiveVideo, setReceiveVideo] = useState<boolean>(false);
   const [receiveAudio, setReceiveAudio] = useState<boolean>(false);
-  const [inputVideoDeviceId, setInputVideoDeviceId] = useState<string>();
-  const [inputAudioDeviceId, setInputAudioDeviceId] = useState<string>();
+  const [sendAudioOptions, setSendAudioOptions] = useState<{
+    inputAudioDeviceId?: string;
+    autoGainControl?: boolean;
+    echoCancellation?: boolean;
+    noiseSuppression?: boolean;
+  }>(
+    localDevice
+      ? {
+          inputAudioDeviceId: localDevice.inputAudioDeviceId || undefined,
+          autoGainControl: localDevice.autoGainControl || false,
+          echoCancellation: localDevice.echoCancellation || false,
+          noiseSuppression: localDevice.noiseSuppression || false,
+        }
+      : {
+          autoGainControl: false,
+          echoCancellation: false,
+          noiseSuppression: false,
+        }
+  );
+  const [inputVideoDeviceId, setInputVideoDeviceId] = useState<string | undefined>(
+    localDevice ? localDevice.inputVideoDeviceId : undefined
+  );
 
   const { ready, consume, produce, stopProducing, stopConsuming } = useMediasoupTransport(
     routerDistUrl
@@ -63,14 +86,16 @@ const MediasoupProvider = (props: {
       setConsumerList((prev) => {
         if (!prev[producer._id]) {
           // CONSUME
-          consume(producer).then((consumer) => {
-            if (isAudioProducer(producer)) {
-              dispatch(allActions.stageActions.client.addAudioConsumer(consumer));
-            } else {
-              dispatch(allActions.stageActions.client.addVideoConsumer(consumer));
-            }
-            return undefined;
-          });
+          consume(producer)
+            .then((consumer) => {
+              if (isAudioProducer(producer)) {
+                dispatch(allActions.stageActions.client.addAudioConsumer(consumer));
+              } else {
+                dispatch(allActions.stageActions.client.addVideoConsumer(consumer));
+              }
+              return undefined;
+            })
+            .catch((error) => reportError(error));
           return {
             ...prev,
             [producer._id]: true,
@@ -94,7 +119,8 @@ const MediasoupProvider = (props: {
               } else {
                 dispatch(allActions.stageActions.client.removeAudioConsumer(consumer._id));
               }
-            });
+            })
+            .catch((error) => reportError(error));
           return omit(prev, consumer.producerId);
         }
         return prev;
@@ -106,32 +132,6 @@ const MediasoupProvider = (props: {
   /** ******************************************
    * VIDEO HANDLING
    */
-  const consumeAllVideos = useCallback(() => {
-    return Promise.all(
-      videoProducers.allIds
-        .map((id) => videoProducers.byId[id])
-        .map((producer) => createConsumer(producer))
-    );
-  }, [videoProducers, createConsumer]);
-  const refreshVideoConsumers = useCallback(() => {
-    const producersWithoutConsumer = videoProducers.allIds
-      .filter((producerId) => !videoConsumers.byProducer[producerId])
-      .map((id) => videoProducers.byId[id]);
-    const consumersWithoutProducer = videoConsumers.allIds
-      .map((id) => videoConsumers.byId[id])
-      .filter((consumer) => !videoProducers.byId[consumer.producerId]);
-    return Promise.all([
-      producersWithoutConsumer.map((producer) => createConsumer(producer)),
-      consumersWithoutProducer.map((consumer) => removeConsumer(consumer, 'video')),
-    ]);
-  }, [videoProducers, videoConsumers, createConsumer, removeConsumer]);
-  const removeAllVideoConsumers = useCallback(() => {
-    return Promise.all(
-      videoConsumers.allIds
-        .map((id) => videoConsumers.byId[id])
-        .map((consumer) => removeConsumer(consumer, 'video'))
-    );
-  }, [videoConsumers, removeConsumer]);
   const shareVideo = useCallback(
     (tracks: MediaStreamTrack[]) => {
       return Promise.all(tracks.map((track) => produce(track))).then((localProducers) =>
@@ -148,63 +148,39 @@ const MediasoupProvider = (props: {
   }, [stopProducing]);
   useEffect(() => {
     if (ready) {
-      debugEffect('ready receiveVideo');
       if (receiveVideo) {
-        consumeAllVideos().catch((error) => err(error));
+        const producersWithoutConsumer = videoProducers.allIds
+          .filter((producerId) => !videoConsumers.byProducer[producerId])
+          .map((id) => videoProducers.byId[id]);
+        const consumersWithoutProducer = videoConsumers.allIds
+          .map((id) => videoConsumers.byId[id])
+          .filter((consumer) => !videoProducers.byId[consumer.producerId]);
+        producersWithoutConsumer.forEach((producer) => createConsumer(producer));
+        consumersWithoutProducer.forEach((consumer) => removeConsumer(consumer, 'audio'));
       } else {
-        removeAllVideoConsumers().catch((error) => err(error));
+        videoConsumers.allIds
+          .map((id) => videoConsumers.byId[id])
+          .forEach((consumer) => removeConsumer(consumer, 'audio'));
       }
     }
-  }, [ready, receiveVideo]);
+    return undefined;
+  }, [ready, receiveVideo, videoConsumers, videoProducers, createConsumer, removeConsumer]);
   useEffect(() => {
     if (ready) {
-      debugEffect('ready videoProducers');
-      if (receiveVideo) {
-        refreshVideoConsumers().catch((error) => err(error));
-      }
-    }
-  }, [ready, videoProducers]);
-  useEffect(() => {
-    if (ready) {
-      debugEffect('ready sendVideo');
+      reportEffect('ready sendVideo');
       if (sendVideo) {
         getVideoTracks(inputVideoDeviceId)
           .then((tracks) => shareVideo(tracks))
-          .catch((error) => err(error));
+          .catch((error) => reportError(error));
         return () => stopSharingVideo();
       }
     }
+    return undefined;
   }, [ready, sendVideo, inputVideoDeviceId, shareVideo, stopSharingVideo]);
 
   /** ******************************************
    * AUDIO HANDLING
    */
-  const consumeAllAudio = useCallback(() => {
-    return Promise.all(
-      audioProducers.allIds
-        .map((id) => audioProducers.byId[id])
-        .map((producer) => createConsumer(producer))
-    );
-  }, [audioProducers, createConsumer]);
-  const refreshAudioConsumers = useCallback(() => {
-    const producersWithoutConsumer = audioProducers.allIds
-      .filter((producerId) => !audioConsumers.byProducer[producerId])
-      .map((id) => audioProducers.byId[id]);
-    const consumersWithoutProducer = audioConsumers.allIds
-      .map((id) => audioConsumers.byId[id])
-      .filter((consumer) => !audioProducers.byId[consumer.producerId]);
-    return Promise.all([
-      producersWithoutConsumer.map((producer) => createConsumer(producer)),
-      consumersWithoutProducer.map((consumer) => removeConsumer(consumer, 'audio')),
-    ]);
-  }, [audioProducers, audioConsumers, createConsumer, removeConsumer]);
-  const removeAllAudioConsumers = useCallback(() => {
-    return Promise.all(
-      audioConsumers.allIds
-        .map((id) => audioConsumers.byId[id])
-        .map((consumer) => removeConsumer(consumer, 'audio'))
-    );
-  }, [audioConsumers, removeConsumer]);
   const shareAudio = useCallback(
     (tracks: MediaStreamTrack[]) => {
       return Promise.all(tracks.map((track) => produce(track))).then((localProducers) =>
@@ -214,7 +190,7 @@ const MediasoupProvider = (props: {
     [produce]
   );
   const stopSharingAudio = useCallback(() => {
-    d('Stop sharing audio');
+    report('Stop sharing audio');
     setLocalAudioProducers((prev) => {
       prev.forEach((localProducer) => stopProducing(localProducer));
       return [];
@@ -223,66 +199,103 @@ const MediasoupProvider = (props: {
   useEffect(() => {
     if (ready) {
       if (receiveAudio) {
-        consumeAllAudio().catch((error) => err(error));
+        const producersWithoutConsumer = audioProducers.allIds
+          .filter((producerId) => !audioConsumers.byProducer[producerId])
+          .map((id) => audioProducers.byId[id]);
+        const consumersWithoutProducer = audioConsumers.allIds
+          .map((id) => audioConsumers.byId[id])
+          .filter((consumer) => !audioProducers.byId[consumer.producerId]);
+
+        report('Add audio consumers');
+        report(producersWithoutConsumer);
+        producersWithoutConsumer.forEach((producer) => createConsumer(producer));
+        report('Remove audio consumers');
+        report(consumersWithoutProducer);
+        consumersWithoutProducer.forEach((consumer) => removeConsumer(consumer, 'audio'));
       } else {
-        removeAllAudioConsumers().catch((error) => err(error));
+        report('remove ALL audio consumers');
+        audioConsumers.allIds
+          .map((id) => audioConsumers.byId[id])
+          .forEach((consumer) => removeConsumer(consumer, 'audio'));
       }
     }
     return undefined;
-  }, [ready, receiveAudio]);
-  useEffect(() => {
-    if (ready) {
-      if (receiveAudio) {
-        refreshAudioConsumers().catch((error) => err(error));
-      }
-    }
-    return undefined;
-  }, [ready, audioProducers]);
+  }, [ready, receiveAudio, audioConsumers, audioProducers, createConsumer, removeConsumer]);
   useEffect(() => {
     if (ready) {
       if (sendAudio) {
-        getAudioTracks(inputAudioDeviceId)
+        getAudioTracks({ ...sendAudioOptions, sampleRate })
           .then((tracks) => shareAudio(tracks))
-          .catch((error) => err(error));
+          .catch((error) => reportError(error));
         return () => stopSharingAudio();
       }
     }
-  }, [ready, sendAudio, inputAudioDeviceId, shareAudio, stopSharingAudio]);
+    return undefined;
+  }, [ready, sendAudio, sendAudioOptions, shareAudio, stopSharingAudio]);
 
   /** *
    * SYNC DEVICE
    */
   useEffect(() => {
     if (ready && localDevice) {
-      if (localDevice.receiveAudio !== receiveAudio) {
-        trace('RECEIVE AUDIO CHANGED');
-        setReceiveAudio(localDevice.receiveAudio);
-      }
-      if (localDevice.receiveVideo !== receiveVideo) {
-        trace('RECEIVE VIDEO CHANGED');
-        setReceiveVideo(localDevice.receiveVideo);
-      }
-      if (localDevice.sendAudio !== sendAudio) {
-        trace('SEND AUDIO CHANGED');
-        setSendAudio(localDevice.sendAudio);
-      }
-      if (localDevice.sendVideo !== sendVideo) {
-        trace('SEND VIDEO CHANGED');
-        setSendVideo(localDevice.sendVideo);
-      }
-      if (localDevice.inputAudioDeviceId !== inputAudioDeviceId) {
-        trace('SEND AUDIO DEVICE ID CHANGED');
-        setInputAudioDeviceId(localDevice.inputAudioDeviceId);
-      }
-      if (localDevice.inputVideoDeviceId !== inputVideoDeviceId) {
-        trace('SEND VIDEO DEVICE ID CHANGED');
-        setInputVideoDeviceId(localDevice.inputVideoDeviceId);
-      }
+      setReceiveAudio((prev) => {
+        if (prev !== localDevice.receiveAudio) {
+          report('RECEIVE AUDIO CHANGED');
+          return localDevice.receiveAudio;
+        }
+        return prev;
+      });
+      setReceiveVideo((prev) => {
+        if (prev !== localDevice.receiveVideo) {
+          report('RECEIVE VIDEO CHANGED');
+          return localDevice.receiveVideo;
+        }
+        return prev;
+      });
+      setSendAudio((prev) => {
+        if (prev !== localDevice.sendAudio) {
+          report('SEND AUDIO CHANGED');
+          return localDevice.sendAudio;
+        }
+        return prev;
+      });
+      setSendVideo((prev) => {
+        if (prev !== localDevice.sendVideo) {
+          report('SEND VIDEO CHANGED');
+          return localDevice.sendVideo;
+        }
+        return prev;
+      });
+      setInputVideoDeviceId((prev) => {
+        if (prev !== localDevice.inputVideoDeviceId) {
+          report('SEND VIDEO DEVICE ID CHANGED');
+          return localDevice.inputVideoDeviceId;
+        }
+        return prev;
+      });
+      setSendAudioOptions((prev) => {
+        report('SEND AUDIO DEVICE CHANGED');
+        if (
+          (localDevice.inputAudioDeviceId !== undefined &&
+            localDevice.inputAudioDeviceId !== prev.inputAudioDeviceId) ||
+          (localDevice.echoCancellation !== undefined &&
+            localDevice.echoCancellation !== prev.echoCancellation) ||
+          (localDevice.autoGainControl !== undefined &&
+            localDevice.autoGainControl !== prev.autoGainControl) ||
+          (localDevice.noiseSuppression !== undefined &&
+            localDevice.noiseSuppression !== prev.noiseSuppression)
+        ) {
+          return {
+            inputAudioDeviceId: localDevice.inputAudioDeviceId || undefined,
+            autoGainControl: localDevice.autoGainControl || false,
+            echoCancellation: localDevice.echoCancellation || false,
+            noiseSuppression: localDevice.noiseSuppression || false,
+          };
+        }
+        return prev;
+      });
     }
   }, [ready, localDevice]);
-  useEffect(() => {
-    trace('MEDIASOUP CHANGED');
-  }, [ready]);
 
   return <>{children}</>;
 };